Add files via upload
This is an Egg catching game made with openGl. This game contains tree Birds which randomly dropping eggs. You have to catch the eggs with a basket the number of egg you catch the higher your level goes.In each level the dropping speed of the eggs increases.you can move the basket with mouse and you can also move the eggs left or right by using button (A&D).

I made the main game with a tutorial on the net and then edited the game and made it as  i wanted.